Move diffing to shared Android module
