+++ shared/src/androidMain/kotlin/dev/johnoreilly/confetti/test/screenshot/ScreenshotTest.kt
@@ -1,0 +1,4 @@
+package dev.johnoreilly.confetti.test.screenshot
+
+abstract class ScreenshotTest {
+}
