+++ wearApp/src/test/kotlin/dev/johnoreilly/confetti/wear/screenshots/ScreenshotTest.kt
@@ -47,6 +47,8 @@ import com.quickbird.snapshot.JUnitFileSnapshotTest
 import com.quickbird.snapshot.Snapshotting
 import com.quickbird.snapshot.fileSnapshotting
 import com.quickbird.snapshot.snapshot
+import dev.johnoreilly.confetti.test.screenshot.bitmapWithTolerance
+import dev.johnoreilly.confetti.test.screenshot.highlightWithRed
 import dev.johnoreilly.confetti.wear.FixedTimeSource
 import dev.johnoreilly.confetti.wear.preview.TestFixtures
 import dev.johnoreilly.confetti.wear.a11y.A11ySnapshotTransformer
